fix: unexpected redirect to signin page when user is singin
