+++ cypress/fixtures/utils.ts
@@ -2,7 +2,8 @@ export const waitGraphql = () =>
   cy.wait('@graphql', {
     timeout: 20_000,
   });
-export const waitSession = () =>
+export const waitSession = () => {
   cy.wait('@session', {
     timeout: 20_000,
   });
+};
