+++ src/pages/api/auth/[...nextauth].ts
@@ -1,4 +1,4 @@
-import NextAuth, { CookiesOptions } from 'next-auth';
+import NextAuth from 'next-auth';
 
 import { getAdminGqlClient } from '$/lib/admin-gql-client';
 import { HASURA_TOKEN_MAX_AGE, SESSION_MAX_AGE } from '$/lib/constants';
@@ -7,6 +7,7 @@ import { nextAuthAdapter } from '$/server/services/auth-adapter';
 import { authProviders } from '$/server/services/auth-providers';
 import { fillUserFields } from '$/server/services/user';
 import { createAuthToken } from '$/server/utilities/create-token';
+import { defaultCookies } from '$/server/utilities/default-cookies';
 import { isENVDev } from '$/server/utilities/env';
 
 export default NextAuth({
@@ -80,71 +81,3 @@ export default NextAuth({
   secret: process.env.HASH_KEY,
   debug: isENVDev,
 });
-
-/**
- * Copy from 'https://github.com/nextauthjs/next-auth/blob/227ff2259f/src/core/lib/cookie.ts' as we can't import it directly
- *
- * Use secure cookies if the site uses HTTPS
- * This being conditional allows cookies to work non-HTTPS development URLs
- * Honour secure cookie option, which sets 'secure' and also adds '__Secure-'
- * prefix, but enable them by default if the site URL is HTTPS; but not for
- * non-HTTPS URLs like http://localhost which are used in development).
- * For more on prefixes see https://googlechrome.github.io/samples/cookie-prefixes/
- *
- */
-export function defaultCookies(useSecureCookies: boolean): CookiesOptions {
-  const cookiePrefix = useSecureCookies ? '__Secure-' : '';
-  // To enable cookies on widgets,
-  // https://stackoverflow.com/questions/45094712/iframe-not-reading-cookies-in-chrome
-  // But we need to set it as `lax` in development
-  const sameSite = isENVDev ? 'lax' : 'none';
-  return {
-    sessionToken: {
-      name: `${cookiePrefix}next-auth.session-token`,
-      options: {
-        httpOnly: true,
-
-        sameSite,
-        path: '/',
-        secure: useSecureCookies,
-      },
-    },
-    callbackUrl: {
-      name: `${cookiePrefix}next-auth.callback-url`,
-      options: {
-        sameSite,
-        path: '/',
-        secure: useSecureCookies,
-      },
-    },
-    csrfToken: {
-      // Default to __Host- for CSRF token for additional protection if using useSecureCookies
-      // NB: The `__Host-` prefix is stricter than the `__Secure-` prefix.
-      name: `${useSecureCookies ? '__Host-' : ''}next-auth.csrf-token`,
-      options: {
-        httpOnly: true,
-        sameSite,
-        path: '/',
-        secure: useSecureCookies,
-      },
-    },
-    pkceCodeVerifier: {
-      name: `${cookiePrefix}next-auth.pkce.code_verifier`,
-      options: {
-        httpOnly: true,
-        sameSite,
-        path: '/',
-        secure: useSecureCookies,
-      },
-    },
-    state: {
-      name: `${cookiePrefix}next-auth.state`,
-      options: {
-        httpOnly: true,
-        sameSite,
-        path: '/',
-        secure: useSecureCookies,
-      },
-    },
-  };
-}
